// frontend/shared/src/graphql/generated/ErpComponentErpProductFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: ErpComponentErpProductFragment
// ====================================================

export interface ErpComponentErpProductFragment {
  __typename: "ErpComponentErpProduct";
  id: number;
  componentId: number;
  productId: number;
  sampleCompanyId: string;
  quantity: number;
}